Move package classes to different java files
